extract field, row, obstacle into their own file
