+++ field_friend/automations/__init__.py
@@ -3,7 +3,8 @@ from .automation_watcher import AutomationWatcher
 from .battery_watcher import BatteryWatcher
 from .coin_collecting import CoinCollecting
 from .coverage_planer import CoveragePlanner
-from .field_provider import Field, FieldObstacle, FieldProvider, Row
+from .field import Field, FieldObstacle, Row
+from .field_provider import FieldProvider
 from .kpi_provider import KpiProvider
 from .mowing import Mowing
 from .path_provider import Path, PathProvider
