+++ field_friend/automations/field_provider.py
@@ -1,65 +1,11 @@
-from dataclasses import dataclass, field
 from typing import Any, Literal, Optional, TypedDict, Union
 
 import rosys
 from geographiclib.geodesic import Geodesic
 from rosys.geometry import Point
-from shapely.geometry import Polygon
 
-from field_friend.navigation import GeoPoint, GeoPointCollection
-
-from .plant import Plant
-
-
-@dataclass(slots=True, kw_only=True)
-class FieldObstacle(GeoPointCollection):
-    pass
-
-
-@dataclass(slots=True, kw_only=True)
-class Row(GeoPointCollection):
-    reverse: bool = False
-    crops: list[Plant] = field(default_factory=list)
-
-    def reversed(self):
-        return Row(
-            id=self.id,
-            name=self.name,
-            points=list(reversed(self.points)),
-        )
-
-    def clear_crops(self):
-        self.crops.clear()
-
-
-@dataclass(slots=True, kw_only=True)
-class Field(GeoPointCollection):
-    reference: Optional[GeoPoint] = None
-    visualized: bool = False
-    obstacles: list[FieldObstacle] = field(default_factory=list)
-    rows: list[Row] = field(default_factory=list)
-
-    @property
-    def outline(self) -> list[rosys.geometry.Point]:
-        assert self.reference is not None
-        return self.cartesian(self.reference)
-
-    @property
-    def outline_as_tuples(self) -> list[tuple[float, float]]:
-        return [p.tuple for p in self.outline]
-
-    def area(self) -> float:
-        outline = self.outline
-        if not outline:
-            return 0.0
-        polygon = Polygon([(p.x, p.y) for p in outline])
-        return polygon.area
-
-    def worked_area(self, worked_rows: int) -> float:
-        worked_area = 0.0
-        if self.area() > 0:
-            worked_area = worked_rows * self.area() / len(self.rows)
-        return worked_area
+from ..navigation import GeoPoint
+from . import Field, FieldObstacle, Row
 
 
 class Active_object(TypedDict):
