+++ field_friend/interface/components/field_planner.py
@@ -34,7 +34,7 @@ class field_planner:
                 with ui.row():
                     self.show_coordinate_type_selection()
                     ui.button('Upload Field', on_click=lambda field_provider=self.field_provider: geodata_picker(field_provider)) \
-                        .tooltip('Upload a file with field boundaries. Supported file formates: KML, XML and Shape') \
+                        .tooltip('Upload a file with field boundaries. Supported file formats: KML, XML and Shape') \
                         .classes('ml-auto').style('display: block; margin-top:auto; margin-bottom: auto;')
                     ui.button('Add field', on_click=self.add_field).tooltip('Add a new field') \
                         .classes('ml-auto').style('display: block; margin-top:auto; margin-bottom: auto;')
